Split resolution queries out of frontend queries
---
