+++ compiler/next/lib/frontend/Makefile.include
@@ -21,8 +21,9 @@ NEXT_FRONTEND_OBJDIR = $(COMPILER_BUILD)/next/lib/frontend
 ALL_SRCS += next/lib/frontend/*.cpp
 
 NEXT_FRONTEND_SRCS =                                \
-           frontend-queries.cpp                     \
            Parser.cpp                               \
+           frontend-queries.cpp                     \
+           resolution-queries.cpp                   \
 
 SRCS = $(NEXT_FRONTEND_SRCS)
